Move codegen Gradle options under the `gradle` subpackage
