+++ mc-java-base/src/main/java/io/spine/tools/mc/java/gradle/codegen/EntityConfig.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.java.codegen;
+package io.spine.tools.mc.java.gradle.codegen;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.GeneratedMessage;
