+++ mc-java-base/src/main/java/io/spine/tools/mc/java/gradle/codegen/ValidationConfig.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.java.codegen;
+package io.spine.tools.mc.java.gradle.codegen;
 
 import io.spine.tools.protoc.Validation;
 import org.gradle.api.Project;
