+++ mc-java-protoc/src/test/java/io/spine/tools/mc/java/protoc/field/FieldGenTest.java
@@ -28,7 +28,7 @@ package io.spine.tools.mc.java.protoc.field;
 
 import com.google.common.testing.NullPointerTester;
 import io.spine.base.SubscribableField;
-import io.spine.tools.mc.java.codegen.FilePatterns;
+import io.spine.tools.mc.java.gradle.codegen.FilePatterns;
 import io.spine.tools.mc.java.protoc.CompilerOutput;
 import io.spine.tools.protoc.CodegenOptions;
 import io.spine.tools.protoc.GenerateFields;
