+++ mc-java-protoc/src/test/java/io/spine/tools/mc/java/protoc/message/InterfaceGenTest.java
@@ -42,7 +42,7 @@ import io.spine.code.java.PackageName;
 import io.spine.tools.java.fs.Directory;
 import io.spine.tools.java.fs.FileName;
 import io.spine.tools.java.fs.SourceFile;
-import io.spine.tools.mc.java.codegen.CodegenOptionsConfig;
+import io.spine.tools.mc.java.gradle.codegen.CodegenOptionsConfig;
 import io.spine.tools.mc.java.protoc.CodeGenerator;
 import io.spine.tools.protoc.CodegenOptions;
 import io.spine.tools.protoc.plugin.message.tests.EveryIsGeneratedProto;
