+++ mc-java-protoc/src/test/java/io/spine/tools/mc/java/protoc/message/InterfacesTest.java
@@ -35,7 +35,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.tools.mc.java.codegen.FilePatterns.fileSuffix;
+import static io.spine.tools.mc.java.gradle.codegen.FilePatterns.fileSuffix;
 import static io.spine.tools.java.code.Names.className;
 
 @DisplayName("`GenerateInterfaces` should")
